diversify files
