--- include/view/tcpublic.h
@@ -1,9 +1,0 @@
-#ifndef VIEW_TCPUBLIC_H
-#define VIEW_TCPUBLIC_H
-
-#include <Python.h>
-#include <view/backport.h>
-
-extern PyTypeObject TCPublicType;
-
-#endif
